# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: StarScreener / TrendingRepo API
  version: 2.0.0
  summary: Trending-repo intelligence platform — canonical profile + discovery + alerts.
  description: |
    Public + authenticated REST surface for [TrendingRepo](https://trendingrepo.com)
    (the StarScreener codebase). The canonical single-repo endpoint is
    `GET /api/repos/{owner}/{name}?v=2`; it returns the full profile (score,
    reasons, mentions, freshness, twitter, npm, ProductHunt, revenue, funding,
    related, prediction, ideas) in one round-trip. Every other `/api/repos/*`
    path returns a subset of the same data, optimised for a narrower caller.

    **Auth summary.** Public reads have no auth. Write endpoints use one of:

      - `Authorization: Bearer <CRON_SECRET>` — scheduled-job / operator routes
      - `Authorization: Bearer <ADMIN_TOKEN>` — admin moderation surface
      - `Authorization: Bearer <USER_TOKEN>` / `x-user-token: <USER_TOKEN>`
        — per-user endpoints (alerts, rules, reactions, ideas)
      - `Cookie: ss_user=<signed-session>` — HMAC-signed browser session
        issued by `POST /api/auth/session`; same userId scope as USER_TOKEN
  contact:
    name: TrendingRepo
    url: https://github.com/0motionguy/starscreener
  license:
    name: MIT
    url: https://github.com/0motionguy/starscreener/blob/main/LICENSE
servers:
  - url: https://trendingrepo.com
    description: Production
  - url: https://starscreener.vercel.app
    description: Vercel canonical
...

...
security: []
...
components:
  securitySchemes:
    cronBearer:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <CRON_SECRET>` — used by scheduled jobs and
        operator-driven data-pipeline endpoints (ingest, persist, recompute,
        cleanup, rebuild, backfill-history, profiles/enrich).
    adminBearer:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <ADMIN_TOKEN>` — browser-facing moderation UI.
        Deliberately NOT interchangeable with `CRON_SECRET` (different blast
        radius).
    userBearer:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <USER_TOKEN>` or `x-user-token: <USER_TOKEN>`
        — per-user endpoints. The `userId` is derived server-side from the
        token; any client-supplied `userId` in body/query is ignored.
    sessionCookie:
      type: apiKey
      in: cookie
      name: ss_user
      description: |
        HMAC-SHA256 signed session cookie issued by `POST /api/auth/session`.
        HttpOnly, SameSite=Lax, 30-day Max-Age. Accepted anywhere `userBearer`
        is accepted.
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Self-serve API key created with `POST /api/keys`. Preferred for
        CLI, MCP, and Data API clients.

  # =========================================================================
  # Reusable primitives
  # =========================================================================
  schemas:
    Error:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
          description: Optional machine-readable code (e.g. `invalid_slug`, `repo_not_found`).
        details:
          type: array
          items: { type: string }
    Timestamp:
      type: string
      format: date-time
      description: ISO-8601 timestamp (UTC).
    FullName:
      type: string
      pattern: "^[A-Za-z0-9._-]+/[A-Za-z0-9._-]+$"
      example: vercel/next.js
    SlugPart:
      type: string
      pattern: "^[A-Za-z0-9._-]+$"
    SocialPlatform:
      type: string
      enum: [reddit, hackernews, bluesky, twitter, devto, github]
...
    MovementStatus:
      type: string
      enum: [hot, breakout, quiet_killer, steady, cooling, dormant]
...
    Repo:
      type: object
      description: |
        Canonical Repo snapshot. Fields are a superset — some (sparklineData,
        crossSignalScore, channelsFiring) are only populated after a
        recompute pass.
      required:
        - id
        - fullName
        - owner
        - name
        - stars
        - language
        - momentumScore
        - movementStatus
      properties:
        id: { type: string, example: vercel--next-js }
        fullName: { $ref: "#/components/schemas/FullName" }
        owner: { type: string }
        name: { type: string }
        ownerAvatarUrl: { type: string, format: uri }
        description: { type: [string, "null"]}
        url: { type: string, format: uri }
        language: { type: [string, "null"]}
        topics:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        categoryId: { type: [string, "null"]}
        stars: { type: integer, minimum: 0 }
        forks: { type: integer, minimum: 0 }
        contributors: { type: integer, minimum: 0 }
        openIssues: { type: integer, minimum: 0 }
        starsDelta24h: { type: integer }
        starsDelta7d: { type: integer }
        starsDelta30d: { type: integer }
        starsDelta24hMissing: { type: boolean }
        forksDelta7d: { type: integer }
        contributorsDelta30d: { type: integer }
        momentumScore: { type: number, minimum: 0, maximum: 100 }
        movementStatus: { $ref: "#/components/schemas/MovementStatus" }
        sparklineData:
          type: array
          items: { type: number }
        rank: { type: integer }
        crossSignalScore: { type: [number, "null"]}
        channelsFiring: { type: [integer, "null"]}
        mentionCount24h: { type: integer, minimum: 0 }
        lastCommitAt: { $ref: "#/components/schemas/Timestamp" }
        lastReleaseAt:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }
        lastReleaseTag: { type: [string, "null"]}
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        archived: { type: boolean }
        deleted: { type: boolean }
        collectionNames:
          type: array
          items: { type: string }
    RepoScore:
      type: object
      description: Score bundle produced by the recompute pipeline.
      properties:
        repoId: { type: string }
        total: { type: number }
        components:
          type: object
          additionalProperties: { type: number }
        computedAt: { $ref: "#/components/schemas/Timestamp" }
    HumanReason:
      type: object
      description: Rendered "why-moving" explainer.
      required: [label, text]
      properties:
        label: { type: string, example: BREAKOUT }
        text: { type: string }
        source: { type: [string, "null"]}
        confidence: { type: [number, "null"]}

    # -----------------------------------------------------------------------
    # Mentions
    # -----------------------------------------------------------------------
    RepoMention:
      type: object
      required: [id, repoId, platform, author, content, url, postedAt]
      properties:
        id: { type: string }
        repoId: { type: string }
        platform: { $ref: "#/components/schemas/SocialPlatform" }
        author: { type: string }
        authorFollowers: { type: [integer, "null"]}
        content: { type: string }
        url: { type: string, format: uri }
        sentiment: { $ref: "#/components/schemas/Sentiment" }
        engagement:
          type: object
          properties:
            likes: { type: integer }
            reposts: { type: integer }
            replies: { type: integer }
            views: { type: integer }
        postedAt: { $ref: "#/components/schemas/Timestamp" }
        discoveredAt: { $ref: "#/components/schemas/Timestamp" }
        isInfluencer: { type: boolean }
        reach: { type: integer }
...
    FreshnessSource:
      type: object
      properties:
        id: { type: string }
        fetchedAt:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }
        ageSeconds: { type: [integer, "null"]}
        cold: { type: boolean }
        stale: { type: boolean }
    FreshnessSnapshot:
      type: object
      required: [fetchedAt, sources]
      properties:
        fetchedAt: { $ref: "#/components/schemas/Timestamp" }
        sources:
          type: array
          items: { $ref: "#/components/schemas/FreshnessSource" }

    # -----------------------------------------------------------------------
    # Funding, revenue, npm, Product Hunt, Twitter
    # -----------------------------------------------------------------------
    FundingEvent:
      type: object
      properties:
        id: { type: string }
        fullName: { $ref: "#/components/schemas/FullName" }
        round: { type: string }
        amountUsd: { type: number }
        announcedAt: { $ref: "#/components/schemas/Timestamp" }
        investors:
          type: array
          items: { type: string }
        source: { type: string }
...
    NpmPackageRow:
      type: object
      properties:
        name: { type: string }
        downloads7d: { type: integer }
        downloads30d: { type: integer }
        lastPublishAt:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }
    Launch:
      type: object
      description: ProductHunt launch record.
      properties:
        slug: { type: string }
        title: { type: string }
        tagline: { type: string }
        votes: { type: integer }
        launchedAt: { $ref: "#/components/schemas/Timestamp" }
    TwitterRepoPanel:
      type: object
      description: |
        Rollup of Twitter signals tied to a repo. Shape defined in
        `src/lib/twitter/types.ts`.
      additionalProperties: true

    # -----------------------------------------------------------------------
    # Related / predictions / ideas
    # -----------------------------------------------------------------------
    RelatedRepoItem:
      type: object
      properties:
        fullName: { $ref: "#/components/schemas/FullName" }
        sharedTopics:
          type: array
          items: { type: string }
        similarityScore: { type: number }
    PredictionItem:
      type: object
      properties:
        id: { type: string }
        fullName: { $ref: "#/components/schemas/FullName" }
        horizonDays: { type: integer, enum: [7, 30, 90] }
        p10: { type: integer }
        p50: { type: integer }
        p90: { type: integer }
        modelVersion: { type: string }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        actualStarsAtHorizon:
          type: [integer, "null"]
        scoredAt:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }
    IdeaItem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        body: { type: string }
        authorHandle: { type: string }
        targetRepoFullName: { $ref: "#/components/schemas/FullName" }
        buildStatus:
          type: string
          enum: [open, building, shipped]
        status:
          type: string
          enum: [published, shipped, archived, pending_moderation, rejected]
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        publishedAt:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }

    # -----------------------------------------------------------------------
    # Canonical profile (the money shape)
    # -----------------------------------------------------------------------
    CanonicalRepoProfileMentions:
      type: object
      required: [recent, nextCursor, countsBySource]
      properties:
        recent:
          type: array
          items: { $ref: "#/components/schemas/RepoMention" }
          description: Newest-first, capped at 50 rows.
        nextCursor:
          type: [string, "null"]
        countsBySource:
          type: object
          additionalProperties: { type: integer, minimum: 0 }
    CanonicalRepoProfileNpm:
      type: object
      required: [packages, dailyDownloads, dependents]
      properties:
        packages:
          type: array
          items: { $ref: "#/components/schemas/NpmPackageRow" }
        dailyDownloads:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              properties:
                date: { type: string, format: date }
                downloads: { type: integer }
        dependents:
          type: object
          additionalProperties:
            type: [integer, "null"]
    CanonicalRepoProfileRevenue:
      type: object
      required: [verified, selfReported, trustmrrClaim]
      properties:
        verified:
          oneOf:
            - { $ref: "#/components/schemas/RevenueOverlay" }
            - { type: "null" }
        selfReported:
          oneOf:
            - { $ref: "#/components/schemas/RevenueOverlay" }
            - { type: "null" }
        trustmrrClaim:
          oneOf:
            - { $ref: "#/components/schemas/RevenueOverlay" }
            - { type: "null" }
    CanonicalRepoProfile:
      type: object
      required:
        - ok
        - fetchedAt
        - v
        - repo
        - mentions
        - freshness
        - npm
        - revenue
        - funding
        - related
      properties:
        ok: { type: boolean, enum: [true] }
        fetchedAt: { $ref: "#/components/schemas/Timestamp" }
        v: { type: integer, enum: [2] }
        repo: { $ref: "#/components/schemas/Repo" }
        score:
          oneOf:
            - { $ref: "#/components/schemas/RepoScore" }
            - { type: "null" }
        reasons:
          type: array
          items: { $ref: "#/components/schemas/HumanReason" }
        mentions: { $ref: "#/components/schemas/CanonicalRepoProfileMentions" }
        freshness: { $ref: "#/components/schemas/FreshnessSnapshot" }
        twitter:
          oneOf:
            - { $ref: "#/components/schemas/TwitterRepoPanel" }
            - { type: "null" }
        npm: { $ref: "#/components/schemas/CanonicalRepoProfileNpm" }
        productHunt:
          oneOf:
            - { $ref: "#/components/schemas/Launch" }
            - { type: "null" }
        revenue: { $ref: "#/components/schemas/CanonicalRepoProfileRevenue" }
        funding:
          type: array
          items: { $ref: "#/components/schemas/FundingEvent" }
        related:
          type: array
          items: { $ref: "#/components/schemas/RelatedRepoItem" }
        prediction:
          oneOf:
            - { $ref: "#/components/schemas/PredictionItem" }
            - { type: "null" }
        ideas:
          type: array
          items: { $ref: "#/components/schemas/IdeaItem" }

    # v1 back-compat shape — kept byte-compatible with the pre-canonical route.
    LegacyRepoProfileV1:
      type: object
      required: [repo, mentions]
      properties:
        repo: { $ref: "#/components/schemas/Repo" }
        score: { type: "null" }
        category: { type: "null" }
        reasons: { type: "null" }
        social:
          type: array
          items: { type: object }
        mentions:
          type: array
          items: { $ref: "#/components/schemas/RepoMention" }
        twitterSignal:
          oneOf:
            - { $ref: "#/components/schemas/TwitterRepoPanel" }
            - { type: "null" }
        whyMoving: { type: array, items: { type: object } }
        relatedRepos:
          type: array
          items: { $ref: "#/components/schemas/RelatedRepoItem" }
        twitterAvailable: { type: boolean }

    # -----------------------------------------------------------------------
    # Alerts
    # -----------------------------------------------------------------------
...
    RepoSubmissionInput:
      type: object
      required: [repo]
      properties:
        repo:
          type: string
          description: Either a GitHub URL or `owner/name`.
        notes: { type: [string, "null"]}
        contact: { type: [string, "null"]}

    # -----------------------------------------------------------------------
    # Reactions + ideas
    # -----------------------------------------------------------------------
...
  parameters:
    OwnerPath:
      name: owner
      in: path
      required: true
      schema: { $ref: "#/components/schemas/SlugPart" }
    NamePath:
      name: name
      in: path
      required: true
      schema: { $ref: "#/components/schemas/SlugPart" }
...
  responses:
    BadRequest:
      description: Invalid parameter or body.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: Internal error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

# ===========================================================================
# Paths
# ===========================================================================
paths:
...
  /api/repos/{owner}/{name}:
    get:
      tags: [Profile]
      summary: Canonical repo profile (v2 default, v1 back-compat).
      description: |
        Default shape (`?v=2`) returns the full `CanonicalRepoProfile`:
        every signal loader output stitched into a single response.

        Pass `?v=1` for the pre-canonical legacy shape (kept byte-compatible
        for MCP tools and CLI consumers pinned to the original contract).
      operationId: getRepoProfile
      parameters:
        - $ref: "#/components/parameters/OwnerPath"
        - $ref: "#/components/parameters/NamePath"
        - name: v
          in: query
          schema: { type: integer, enum: [1, 2], default: 2 }
      responses:
        "200":
          description: Full profile.
          headers:
            Cache-Control:
              schema: { type: string }
              example: public, s-maxage=30, stale-while-revalidate=60
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/CanonicalRepoProfile" }
                  - { $ref: "#/components/schemas/LegacyRepoProfileV1" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /api/repos/{owner}/{name}/aiso:
    get:
      tags: [Profile, Evidence]
      summary: AISO scan status for a repo.
      operationId: getAisoStatus
      parameters:
        - $ref: "#/components/parameters/OwnerPath"
        - $ref: "#/components/parameters/NamePath"
      responses:
        "200":
          description: Current scan status.
          content:
            application/json:
              schema:
                type: object
                required: [ok, status]
                properties:
                  ok: { type: boolean, enum: [true] }
                  status:
                    type: string
                    enum: [scanned, queued, rate_limited, failed, none]
                  score: { type: [number, "null"]}
                  tier: { type: [string, "null"]}
                  dimensions:
                    type: array
                    items: { type: object, additionalProperties: true }
                  topDimensions:
                    type: array
                    items: { type: object, additionalProperties: true }
                  lastScanAt:
                    oneOf:
                      - { $ref: "#/components/schemas/Timestamp" }
                      - { type: "null" }
                  signals:
                    type: [object, "null"]
                    additionalProperties: true
                  engineCitations:
                    type: array
                    items:
                      type: object
                      properties:
                        engine: { type: string }
                        cited: { type: boolean }
                        brandMentioned: { type: boolean }
                        position: { type: [integer, "null"]}
                  resultUrl: { type: [string, "null"], format: uri }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Profile, Evidence]
      summary: Enqueue an AISO rescan for this repo.
      description: |
        Rate-limited to 1 request per 60s per IP. Appends a row to the
        rescan queue (no inline scan work).
      operationId: enqueueAisoRescan
      parameters:
        - $ref: "#/components/parameters/OwnerPath"
        - $ref: "#/components/parameters/NamePath"
      responses:
        "200":
          description: Queued.
          content:
            application/json:
              schema:
                type: object
                required: [ok, status, queuedAt]
                properties:
                  ok: { type: boolean, enum: [true] }
                  status: { type: string, enum: [queued] }
                  queuedAt: { $ref: "#/components/schemas/Timestamp" }
                  queuePath: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/ServerError" }

  # -------------------------------------------------------------------------
  # Discovery
  # -------------------------------------------------------------------------
...
  /api/repo-submissions:
    get:
      tags: [Submissions]
      summary: List queued repo submissions + summary.
      operationId: listRepoSubmissions
      responses:
        "200":
          description: Queue + recent rows.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    post:
      tags: [Submissions]
      summary: Submit a repo for intake.
      operationId: submitRepo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RepoSubmissionInput" }
      responses:
        "200":
          description: Queued + maybe intake triggered.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400": { $ref: "#/components/responses/BadRequest" }

  # -------------------------------------------------------------------------
  # Predictions / calibration / revenue-estimator / profile
  # -------------------------------------------------------------------------
...